HomeWork-7
NEW: Added file modules.
